chore(migrate): glint comment bug repro for nested long line
In a nested helper invocation, glint-expect-error comments are
injected at incorrect positions.

This is because a formatter is run and attempts to single-line the
helper invocation.

If the nested helper invocation is short enough it will move the
comment up.

If it's too long, it will keep the enumerated args, and keep the bad
comments in place.